Status and mobile UI added

+++ lib/manager/models/contract_md.dart
@@ -24,8 +24,8 @@ class ContractMd {
   bool? AHEonYS;
   int? initHolidays;
   String? jobDescription;
-  int? salaryPH;
-  int? salaryOT;
+  double? salaryPH;
+  double? salaryOT;
   String? jobTitleId;
 
   @override
